requirements.txt
-e .
attrs~=20.2.0
aws-cdk.aws-dynamodb~=1.71.0
boto3~=1.16.8
pip~=20.2.4
parallelagram~=0.0.1
py~=1.9.0
pytest~=6.1.2
setuptools~=50.3.2
